rename headers to have hpp extension
