+++ code/src/stf/alg/transform.hpp
@@ -2,7 +2,7 @@
 
 #include <cmath>
 
-#include "../math/vector.h"
+#include "../math/vector.hpp"
 
 namespace stf {
 namespace alg {
