+++ code/src/stf/math/constants.hpp
