+++ code/tests/alg/interpolation_tests.cpp
@@ -2,7 +2,7 @@
 
 #include <gtest/gtest.h>
 
-#include <stf/types.h>
+#include <stf/types.hpp>
 
 namespace stf {
 namespace alg {
