+++ code/tests/alg/transform_tests.cpp
@@ -2,10 +2,10 @@
 
 #include <gtest/gtest.h>
 
-#include <stf/types.h>
-#include <stf/alg/transform.h>
+#include <stf/types.hpp>
+#include <stf/alg/transform.hpp>
 
-#include "alg/scaffolding/transform.h"
+#include "alg/scaffolding/transform.hpp"
 
 namespace stf {
 namespace alg {
