+++ code/tests/math/scaffolding/vector.hpp
@@ -2,7 +2,7 @@
 
 #include <gtest/gtest.h>
 
-#include <stf/math/vector.h>
+#include <stf/math/vector.hpp>
 
 namespace stf {
 namespace math {
